refactored and moved npm fetch to util file.
